Modified TC 115 to pass Expat/SAX-based processors.
In the implementation case where RDFa processors do not have access
to the raw byte stream, they would inevitably not see encoded XML
entities. Therefore, the byte stream &lt;strong&gt; would be transformed
to the "<strong>" bytestream by the XML processor before being handed to
the RDFa processor. This would result in the RDFa Processor assuming the
beginning of an XMLLiteral. Since the purpose of TC 115 was to test
entity translation, datatype="" was added to ensure that Expat/SAX-based
parsers would also pass this test if they performed proper entity
translation.