Remove xhtml1 from test 0109, which is defined for xhtml1, html4, and html5 for both RDFa 1.0 and 1.1. It is there to ensure that @xml:base is ignored. Shane points out that XHTML 1.1 specifically prohibits the use of @xml:base, but Alex responds that that would be an XHTML conformance issue, not an RDFa conformance issue.
Remove html4 and html5 from test 0256, which uses @lang and @xml:lang on the same element, which is defined for xhtml1, html4, html5, and xhtml5. HTML5+RDFa 1.1 specifically talks about @lang and @xml:lang on the same element and that @xml:lang takes precedence. HTML5 prohibits the use of @lang in the XML namespace in HTML documents, only in XML documents. For compatibility purposes, the @xml:lang attribute in no namespace MAY be used, but only if an @lang attribute is also specified, and they must have the same value. Furthermore, even in XHTML5, the values of @xml:lang and @lang, if both present MUST have the same value. The value of this test for RDFa is questionable, as it tests something which is illegal, at least in (X)HTML5.